requirements.txt
Django==1.11
django_rq==0.9.0
django-jsonfield==1.1.0